@b-deam 2023-02-24
Thanks to @danielmitterdorfer's guidance we were able to resolve the logging 'bug'. Because the actor system effectively double forks itself, if module imports are eager (i.e. at the top of a module), then any attempt to configure them after the fact can be ignored.

Commit https://github.com/elastic/rally/pull/1669/commits/56506610ae37633843ed8064f0c9e376f6d1a4d5 was the fix.

In other news, I ran through the various `tsdb` challenges (with the logging bug present), and again, the results appear consistent:

<details>

<summary> tsdb-standard </summary>

```
baseline: 4ebf4272-bb43-4c52-8246-0de2d2a024f6
contender: dadc1c0e-633f-4d6f-b0fa-de8900e7da7d

$ esrally compare --baseline 4ebf4272-bb43-4c52-8246-0de2d2a024f6 --contender dadc1c0e-633f-4d6f-b0fa-de8900e7da7d

    ____        ____
   / __ \____ _/ / /_  __
  / /_/ / __ `/ / / / / /
 / _, _/ /_/ / / / /_/ /
/_/ |_|\__,_/_/_/\__, /
                /____/


Comparing baseline
  Race ID: 4ebf4272-bb43-4c52-8246-0de2d2a024f6
  Race timestamp: 2022-12-25 20:00:00
  Challenge: append-no-conflicts
  Car: external
  User tags: benchmark-name=tsdb, benchmark-type=rally-es-client-upgrade, created-by=esbench, division=engineering, env-id=c9166eb8-127c-4b9d-8ff8-463af55784d6, git-username=Brad Deam, host-username=bradleydeam, name=tsdb-standard, org=elasticsearch, project=esbench-nightly, setup=esbench, team=performance

with contender
  Race ID: dadc1c0e-633f-4d6f-b0fa-de8900e7da7d
  Race timestamp: 2022-12-25 20:00:00
  Challenge: append-no-conflicts
  Car: external
  User tags: benchmark-name=tsdb, benchmark-type=rally-es-client-upgrade, created-by=esbench, division=engineering, env-id=60ab20a9-5dab-45eb-8879-b0d2ec789121, git-username=Brad Deam, host-username=bradleydeam, name=tsdb-standard, org=elasticsearch, project=esbench-nightly, setup=esbench, team=performance

------------------------------------------------------
    _______             __   _____
   / ____(_)___  ____ _/ /  / ___/_________  ________
  / /_  / / __ \/ __ `/ /   \__ \/ ___/ __ \/ ___/ _ \
 / __/ / / / / / /_/ / /   ___/ / /__/ /_/ / /  /  __/
/_/   /_/_/ /_/\__,_/_/   /____/\___/\____/_/   \___/
------------------------------------------------------
            
|                                                        Metric |                    Task |        Baseline |       Contender |        Diff |   Unit |   Diff % |
|--------------------------------------------------------------:|------------------------:|----------------:|----------------:|------------:|-------:|---------:|
|                    Cumulative indexing time of primary shards |                         |   326.167       |   333.913       |     7.74572 |    min |   +2.37% |
|             Min cumulative indexing time across primary shard |                         |   326.167       |   333.913       |     7.74572 |    min |   +2.37% |
|          Median cumulative indexing time across primary shard |                         |   326.167       |   333.913       |     7.74572 |    min |   +2.37% |
|             Max cumulative indexing time across primary shard |                         |   326.167       |   333.913       |     7.74572 |    min |   +2.37% |
|           Cumulative indexing throttle time of primary shards |                         |     0.0976      |     0.00895     |    -0.08865 |    min |  -90.83% |
|    Min cumulative indexing throttle time across primary shard |                         |     0.0976      |     0.00895     |    -0.08865 |    min |  -90.83% |
| Median cumulative indexing throttle time across primary shard |                         |     0.0976      |     0.00895     |    -0.08865 |    min |  -90.83% |
|    Max cumulative indexing throttle time across primary shard |                         |     0.0976      |     0.00895     |    -0.08865 |    min |  -90.83% |
|                       Cumulative merge time of primary shards |                         |   169.931       |   168.96        |    -0.97088 |    min |   -0.57% |
|                      Cumulative merge count of primary shards |                         |  1061           |  1063           |     2       |        |   +0.19% |
|                Min cumulative merge time across primary shard |                         |   169.931       |   168.96        |    -0.97088 |    min |   -0.57% |
|             Median cumulative merge time across primary shard |                         |   169.931       |   168.96        |    -0.97088 |    min |   -0.57% |
|                Max cumulative merge time across primary shard |                         |   169.931       |   168.96        |    -0.97088 |    min |   -0.57% |
|              Cumulative merge throttle time of primary shards |                         |     1.05617     |     2.59102     |     1.53485 |    min | +145.32% |
|       Min cumulative merge throttle time across primary shard |                         |     1.05617     |     2.59102     |     1.53485 |    min | +145.32% |
|    Median cumulative merge throttle time across primary shard |                         |     1.05617     |     2.59102     |     1.53485 |    min | +145.32% |
|       Max cumulative merge throttle time across primary shard |                         |     1.05617     |     2.59102     |     1.53485 |    min | +145.32% |
|                     Cumulative refresh time of primary shards |                         |     1.32082     |     1.3251      |     0.00428 |    min |   +0.32% |
|                    Cumulative refresh count of primary shards |                         |   330           |   330           |     0       |        |    0.00% |
|              Min cumulative refresh time across primary shard |                         |     1.32082     |     1.3251      |     0.00428 |    min |   +0.32% |
|           Median cumulative refresh time across primary shard |                         |     1.32082     |     1.3251      |     0.00428 |    min |   +0.32% |
|              Max cumulative refresh time across primary shard |                         |     1.32082     |     1.3251      |     0.00428 |    min |   +0.32% |
|                       Cumulative flush time of primary shards |                         |     5.13445     |     5.2056      |     0.07115 |    min |   +1.39% |
|                      Cumulative flush count of primary shards |                         |   302           |   304           |     2       |        |   +0.66% |
|                Min cumulative flush time across primary shard |                         |     5.13445     |     5.2056      |     0.07115 |    min |   +1.39% |
|             Median cumulative flush time across primary shard |                         |     5.13445     |     5.2056      |     0.07115 |    min |   +1.39% |
|                Max cumulative flush time across primary shard |                         |     5.13445     |     5.2056      |     0.07115 |    min |   +1.39% |
|                                       Total Young Gen GC time |                         |     2.909       |     2.917       |     0.008   |      s |   +0.28% |
|                                      Total Young Gen GC count |                         |   214           |   215           |     1       |        |   +0.47% |
|                                         Total Old Gen GC time |                         |     0           |     0           |     0       |      s |    0.00% |
|                                        Total Old Gen GC count |                         |     0           |     0           |     0       |        |    0.00% |
|                                                    Store size |                         |    31.179       |    31.2005      |     0.02155 |     GB |   +0.07% |
|                                                 Translog size |                         |     5.12227e-08 |     5.12227e-08 |     0       |     GB |    0.00% |
|                                        Heap used for segments |                         |     0           |     0           |     0       |     MB |    0.00% |
|                                      Heap used for doc values |                         |     0           |     0           |     0       |     MB |    0.00% |
|                                           Heap used for terms |                         |     0           |     0           |     0       |     MB |    0.00% |
|                                           Heap used for norms |                         |     0           |     0           |     0       |     MB |    0.00% |
|                                          Heap used for points |                         |     0           |     0           |     0       |     MB |    0.00% |
|                                   Heap used for stored fields |                         |     0           |     0           |     0       |     MB |    0.00% |
|                                                 Segment count |                         |     1           |     1           |     0       |        |    0.00% |
|                                   Total Ingest Pipeline count |                         |     0           |     0           |     0       |        |    0.00% |
|                                    Total Ingest Pipeline time |                         |     0           |     0           |     0       |     ms |    0.00% |
|                                  Total Ingest Pipeline failed |                         |     0           |     0           |     0       |        |    0.00% |
|                                                Min Throughput |                   index | 84257.5         | 81953.1         | -2304.43    | docs/s |   -2.73% |
|                                               Mean Throughput |                   index | 85948.5         | 83537.7         | -2410.77    | docs/s |   -2.80% |
|                                             Median Throughput |                   index | 85609.5         | 83422.5         | -2186.91    | docs/s |   -2.55% |
|                                                Max Throughput |                   index | 88144.4         | 85321.7         | -2822.64    | docs/s |   -3.20% |
|                                       50th percentile latency |                   index |   880.482       |   904.476       |    23.9932  |     ms |   +2.73% |
|                                       90th percentile latency |                   index |  1263.62        |  1280.26        |    16.6337  |     ms |   +1.32% |
|                                       99th percentile latency |                   index |  1617.91        |  1619.46        |     1.54429 |     ms |   +0.10% |
|                                     99.9th percentile latency |                   index |  2145.17        |  1838.13        |  -307.037   |     ms |  -14.31% |
|                                    99.99th percentile latency |                   index |  2768.97        |  2172.77        |  -596.195   |     ms |  -21.53% |
|                                      100th percentile latency |                   index |  2888.27        |  2409.64        |  -478.629   |     ms |  -16.57% |
|                                  50th percentile service time |                   index |   880.62        |   904.499       |    23.8796  |     ms |   +2.71% |
|                                  90th percentile service time |                   index |  1263.47        |  1280.37        |    16.8995  |     ms |   +1.34% |
|                                  99th percentile service time |                   index |  1618.16        |  1619.6         |     1.43712 |     ms |   +0.09% |
|                                99.9th percentile service time |                   index |  2145.17        |  1838.13        |  -307.037   |     ms |  -14.31% |
|                               99.99th percentile service time |                   index |  2768.97        |  2172.77        |  -596.195   |     ms |  -21.53% |
|                                 100th percentile service time |                   index |  2888.27        |  2409.64        |  -478.629   |     ms |  -16.57% |
|                                                    error rate |                   index |     0           |     0           |     0       |      % |    0.00% |
|                                                Min Throughput |                 default |   309.33        |   238.371       |   -70.9588  |  ops/s |  -22.94% |
|                                               Mean Throughput |                 default |   309.33        |   238.371       |   -70.9588  |  ops/s |  -22.94% |
|                                             Median Throughput |                 default |   309.33        |   238.371       |   -70.9588  |  ops/s |  -22.94% |
|                                                Max Throughput |                 default |   309.33        |   238.371       |   -70.9588  |  ops/s |  -22.94% |
|                                       50th percentile latency |                 default |     2.52969     |     2.98391     |     0.45422 |     ms |  +17.96% |
|                                       90th percentile latency |                 default |     2.94294     |     3.31405     |     0.37112 |     ms |  +12.61% |
|                                       99th percentile latency |                 default |     4.14364     |     4.83318     |     0.68954 |     ms |  +16.64% |
|                                      100th percentile latency |                 default |     4.68847     |     5.55966     |     0.87119 |     ms |  +18.58% |
|                                  50th percentile service time |                 default |     2.52969     |     2.98391     |     0.45422 |     ms |  +17.96% |
|                                  90th percentile service time |                 default |     2.94294     |     3.31405     |     0.37112 |     ms |  +12.61% |
|                                  99th percentile service time |                 default |     4.14364     |     4.83318     |     0.68954 |     ms |  +16.64% |
|                                 100th percentile service time |                 default |     4.68847     |     5.55966     |     0.87119 |     ms |  +18.58% |
|                                                    error rate |                 default |     0           |     0           |     0       |      % |    0.00% |
|                                                Min Throughput |              default_1k |    81.1999      |    71.6301      |    -9.56984 |  ops/s |  -11.79% |
|                                               Mean Throughput |              default_1k |    81.1999      |    71.6301      |    -9.56984 |  ops/s |  -11.79% |
|                                             Median Throughput |              default_1k |    81.1999      |    71.6301      |    -9.56984 |  ops/s |  -11.79% |
|                                                Max Throughput |              default_1k |    81.1999      |    71.6301      |    -9.56984 |  ops/s |  -11.79% |
|                                       50th percentile latency |              default_1k |     8.32006     |     8.96284     |     0.64279 |     ms |   +7.73% |
|                                       90th percentile latency |              default_1k |     8.71812     |     9.55749     |     0.83937 |     ms |   +9.63% |
|                                       99th percentile latency |              default_1k |    10.7995      |    11.1296      |     0.33012 |     ms |   +3.06% |
|                                      100th percentile latency |              default_1k |    10.8964      |    11.4962      |     0.59979 |     ms |   +5.50% |
|                                  50th percentile service time |              default_1k |     8.32006     |     8.96284     |     0.64279 |     ms |   +7.73% |
|                                  90th percentile service time |              default_1k |     8.71812     |     9.55749     |     0.83937 |     ms |   +9.63% |
|                                  99th percentile service time |              default_1k |    10.7995      |    11.1296      |     0.33012 |     ms |   +3.06% |
|                                 100th percentile service time |              default_1k |    10.8964      |    11.4962      |     0.59979 |     ms |   +5.50% |
|                                                    error rate |              default_1k |     0           |     0           |     0       |      % |    0.00% |
|                                                Min Throughput | date-histo-entire-range |   326.476       |   240.238       |   -86.2378  |  ops/s |  -26.41% |
|                                               Mean Throughput | date-histo-entire-range |   326.476       |   240.238       |   -86.2378  |  ops/s |  -26.41% |
|                                             Median Throughput | date-histo-entire-range |   326.476       |   240.238       |   -86.2378  |  ops/s |  -26.41% |
|                                                Max Throughput | date-histo-entire-range |   326.476       |   240.238       |   -86.2378  |  ops/s |  -26.41% |
|                                       50th percentile latency | date-histo-entire-range |     1.89445     |     2.33147     |     0.43701 |     ms |  +23.07% |
|                                       90th percentile latency | date-histo-entire-range |     2.25338     |     2.69167     |     0.43829 |     ms |  +19.45% |
|                                       99th percentile latency | date-histo-entire-range |     5.67756     |     3.53429     |    -2.14327 |     ms |  -37.75% |
|                                      100th percentile latency | date-histo-entire-range |     7.50967     |     3.78365     |    -3.72602 |     ms |  -49.62% |
|                                  50th percentile service time | date-histo-entire-range |     1.89445     |     2.33147     |     0.43701 |     ms |  +23.07% |
|                                  90th percentile service time | date-histo-entire-range |     2.25338     |     2.69167     |     0.43829 |     ms |  +19.45% |
|                                  99th percentile service time | date-histo-entire-range |     5.67756     |     3.53429     |    -2.14327 |     ms |  -37.75% |
|                                 100th percentile service time | date-histo-entire-range |     7.50967     |     3.78365     |    -3.72602 |     ms |  -49.62% |
|                                                    error rate | date-histo-entire-range |     0           |     0           |     0       |      % |    0.00% |


-------------------------------
[INFO] SUCCESS (took 0 seconds)
-------------------------------
```

</details>

<details>

<summary> tsdb-standard-synthetic-source </summary>

```
baseline: f199a835-d564-4eb1-848f-dc95ea68a17f
contender: 8dc5fa0b-5c28-42b8-8300-36d038a7e0b7

$ esrally compare --baseline f199a835-d564-4eb1-848f-dc95ea68a17f --contender 8dc5fa0b-5c28-42b8-8300-36d038a7e0b7

    ____        ____
   / __ \____ _/ / /_  __
  / /_/ / __ `/ / / / / /
 / _, _/ /_/ / / / /_/ /
/_/ |_|\__,_/_/_/\__, /
                /____/


Comparing baseline
  Race ID: f199a835-d564-4eb1-848f-dc95ea68a17f
  Race timestamp: 2022-12-25 20:00:00
  Challenge: append-no-conflicts
  Car: external
  User tags: benchmark-name=tsdb, benchmark-type=rally-es-client-upgrade, created-by=esbench, division=engineering, env-id=c9166eb8-127c-4b9d-8ff8-463af55784d6, git-username=Brad Deam, host-username=bradleydeam, name=tsdb-standard-synthetic-source, org=elasticsearch, project=esbench-nightly, setup=esbench, team=performance

with contender
  Race ID: 8dc5fa0b-5c28-42b8-8300-36d038a7e0b7
  Race timestamp: 2022-12-25 20:00:00
  Challenge: append-no-conflicts
  Car: external
…